Use pip to install/package.
Closes #2. For reasons unknown to me, not using pip to install/package a
noarch python module can sometimes leave the scripts in the package
without a execute bit. When these packages are installed, the script
cannot be executed without prefixing python:
```
python /path/to/bin/pyxbgen
```

This commit changes the meta.yaml file to use pip to install/package
pyxb. It also adds pip and setuptools to the requirements for building.